Complete main function
We finished the main function implementation. We now support two output
formats: XML and debug printing. The former is used by the test cases,
and the latter is Rust debug formatting with pretty-print, which shows
the whole structure and data of the built AST. We also support writing
the output to a specified file instead of just stdout.